Changed killscript into internal command with script
Instead of having a killscript option for a service configuration,
I decided it needed to be more like a shell function or external
command.  The implementation will now be an internal function which
can be executed over the control socket, and I added an external
helper script to make the call with proper error handling and
diagnostics.

Since the script can find the control socket and service name from
its environment variables, a bash script calling it can be as simple
as

  descall-kill SIGTERM 10 SIGKILL 5

which makes it easy to string into a one-liner:

  action stop:
    echo terminate | socat UNIX:/path/to/service/socket
    && descall-kill SIGTERM 10 SIGKILL 5
    && rm -rf /path/to/cache/dir

if you have a complicated service shutdown sequence.