Reworked the structure of the config file
Service start action no longer becomes the service, instead there
is a 'run' attribute for the service itself, and the start action
calls to an internal method to do the exec, and then returns a
code indicating success or failure.

Changed 'set-mode' to 'goal', and now actions only apply a goal
if they are called directly by the user.

Run values can now be maps and specify internal calls.

Killscript can now happen without forking at all, which will be
much more efficient.

Control handles are no longer given to actions by default.

Reserved some attributes for the yet-to-be-defined permission
system.

Specified 'check' as a standard action, which defaults to a check
if the service has been running for 2 seconds.